+++ test/Jenkins/Rest/InternalSpec.hs
@@ -4,14 +4,13 @@ module Jenkins.Rest.InternalSpec (spec) where
 import           Control.Lens
 import           Control.Exception (throwIO)
 import           Control.Exception.Lens (throwingM, _IOException)
-import           Control.Monad.IO.Class (liftIO)
 import           Network.HTTP.Conduit (HttpException)
 import           Network.HTTP.Types (Status(..))
 import           Test.Hspec.Lens
 import           System.IO.Error
 import           System.IO.Error.Lens (errorType, _NoSuchThing)
 
-import qualified Jenkins.Rest as Helper
+import           Jenkins.Rest
 import           Jenkins.Rest.Internal
 import           Network.HTTP.Conduit.Lens (_StatusCodeException, _InvalidUrlException, _TooManyRetries)
 
@@ -20,20 +19,20 @@ _JenkinsException = iso (\(JenkinsHttpException e) -> e) JenkinsHttpException
 
 spec :: Spec
 spec = do
-  let raiseHttp, raiseIO :: Jenkins a
+  let raiseHttp, raiseIO :: JenkinsT IO a
       raiseHttp = liftIO (throwingM _TooManyRetries ())
       raiseIO   = liftIO (throwIO (mkIOError doesNotExistErrorType "foo" Nothing Nothing))
 
   describe "runJenkins" $ do
     it "wraps uncatched 'HttpException' exceptions from the queries in 'Error'" $
-      runJenkins defaultConnectInfo (Helper.get Helper.plain "hi")
+      runJenkins defaultConnectInfo (get plain "hi")
      `shouldPerform`
       Status 404 ""
      `through`
       _Error._JenkinsException._StatusCodeException._1
 
     it "wraps uncatched 'HttpException' exceptions from the URL parsing in 'Error'" $
-      runJenkins (defaultConnectInfo & jenkinsUrl .~ "foo") (Helper.get Helper.plain "hi")
+      runJenkins (defaultConnectInfo & jenkinsUrl .~ "foo") (get plain "hi")
      `shouldPerform`
       ("foo", "Invalid URL")
      `through`
@@ -41,7 +40,7 @@ spec = do
 
     it "can catch 'HttpException' exceptions related from the queries" $
       runJenkins defaultConnectInfo
-        (liftJ (Or (Helper.get Helper.plain "hi" >> return 4) (return 7)))
+        (liftJ (Or (get plain "hi" >> return 4) (return 7)))
      `shouldPerform`
       7
      `through`